requirements.txt
# Core Application
pydantic
pydantic-settings

# Web Scraping & Details
marktplaats
requests
beautifulsoup4

# LLM Clients
together
openai
azure-identity

# Data Handling & Storage
pandas
google-cloud-firestore
google-cloud-storage

# OCR
google-cloud-vision

# Web Generation
Jinja2

# Testing
pytest
pytest-mock

# Utility
python-dotenv